requirements.txt
rpi.gpio~=0.7.0
aiohttp~=3.7.4
cchardet~=2.1.6
aiodns~=2.0.0
transitions~=0.8.2
pygraphviz~=1.6
spidev~=3.5
prometheus_client~=0.8.0
python-dateutil~=2.8.1
pandas~=1.2.3
odfpy~=1.4.1
attrs~=20.3.0
